much better logic for setting `game.installed`
Leveraged existing backend function to check whether a path exists (our game in this case) in a platform-independent way so it works regardless of path separator.

This now means the config files button, installed tab and open folder button all react properly if the game is moved.

TODO: Check for game executable in addition to dir.