Clicking install/uninstall now links/unlinks the mod and every dependency to the loader's mod path.
I'm not sure I like how messy this implementation is and probably will be somewhat refactored in future. This may be just a symptom of not having ecosystem.json fully implemented right now.
NOTE: Mods that have their own BepInEx folder aren't handled and won't work in most cases when running the game.